refactor: extract file watcher into seperate file
